update state types

+++ app/models/state.py
@@ -2,5 +2,5 @@ from pydantic import BaseModel
 
 class State(BaseModel):
   avg_score: float
-  avg_res_time: int
+  avg_res_time: float
   current_difficulty: int
